Design improvements
