--- app/src/main/res/layout/fragment_about.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginTop="10dp"
-    android:orientation="vertical">
-
-    <ImageView
-        android:layout_width="match_parent"
-        android:layout_height="140dp"
-        android:layout_marginBottom="20dp"
-        android:importantForAccessibility="no"
-        android:scaleType="fitCenter"
-        android:src="@drawable/icon_grayscale">
-    </ImageView>
-
-    <TextView
-        android:id="@+id/ApplicationVersion"
-        android:layout_marginStart="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textSize="16sp"
-        android:textStyle="normal"
-        android:ellipsize="marquee">
-    </TextView>
-
-    <TextView
-        android:id="@+id/RootfsVersion"
-        android:layout_marginStart="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textSize="16sp"
-        android:textStyle="normal"
-        android:ellipsize="marquee">
-    </TextView>
-
-    <TextView
-        android:id="@+id/Box64Version"
-        android:layout_marginStart="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textSize="16sp"
-        android:textStyle="normal"
-        android:ellipsize="marquee">
-    </TextView>
-
-    <TextView
-        android:id="@+id/WineVersion"
-        android:layout_marginStart="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textSize="16sp"
-        android:textStyle="normal"
-        android:ellipsize="marquee">
-    </TextView>
-
-</LinearLayout>
