+++ app/src/main/res/layout/fragment_home.xml
@@ -2,17 +2,107 @@
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:background="@android:color/transparent">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerViewGame"
-        android:layout_width="wrap_content"
+    <com.google.android.material.card.MaterialCardView
+        android:id="@+id/info_card"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="10dp"
-        android:layout_marginTop="10dp"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-    </androidx.recyclerview.widget.RecyclerView>
+        android:layout_marginHorizontal="24dp"
+        android:layout_gravity="center"
+        app:cardBackgroundColor="@android:color/transparent"
+        app:cardCornerRadius="16dp"
+        app:cardElevation="0dp"
+        app:strokeColor="?attr/colorOutline"
+        app:strokeWidth="1dp">
 
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:padding="20dp">
+
+            <ImageView
+                android:layout_width="64dp"
+                android:layout_height="64dp"
+                android:layout_gravity="center"
+                android:layout_marginBottom="20dp"
+                android:contentDescription="@null"
+                android:scaleType="centerInside"
+                android:src="@drawable/icon_grayscale" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/ApplicationVersionTitle"
+                style="@style/TextAppearance.Material3.TitleSmall"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Application Version"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/ApplicationVersion"
+                style="@style/TextAppearance.Material3.BodyMedium"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:text="1.0.0"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/RootfsVersionTitle"
+                style="@style/TextAppearance.Material3.TitleSmall"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="24dp"
+                android:text="RootFS Version"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/RootfsVersion"
+                style="@style/TextAppearance.Material3.BodyMedium"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:text="RootFS 1.1.0"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/Box64VersionTitle"
+                style="@style/TextAppearance.Material3.TitleSmall"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="24dp"
+                android:text="Box64 Version"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/Box64Version"
+                style="@style/TextAppearance.Material3.BodyMedium"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:text="1.2.3"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/WineVersionTitle"
+                style="@style/TextAppearance.Material3.TitleSmall"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="24dp"
+                android:text="Wine Version"
+                android:textColor="?attr/colorOnBackground" />
+
+            <com.google.android.material.textview.MaterialTextView
+                android:id="@+id/WineVersion"
+                style="@style/TextAppearance.Material3.BodyMedium"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:text="7.8"
+                android:textColor="?attr/colorOnBackground" />
+
+        </LinearLayout>
+    </com.google.android.material.card.MaterialCardView>
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
