+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -8,6 +8,13 @@
         android:title="@string/home_title">
     </item>
     <item
+        android:id="@+id/nav_shortcuts"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:icon="@drawable/ic_shortcut"
+        android:title="@string/shortcuts_title">
+    </item>
+    <item
         android:id="@+id/nav_settings"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -20,13 +27,6 @@
         android:layout_height="wrap_content"
         android:icon="@drawable/ic_folder"
         android:title="@string/file_manager_title">
-    </item>
-    <item
-        android:id="@+id/nav_about_micewine"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:icon="@drawable/ic_info_outline"
-        android:title="@string/about_title">
     </item>
 </menu>
