+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -34,7 +34,7 @@
     <string name="log_description">查看 MiceWine 日志</string>
     <string name="clear_log">清除</string>
     <string name="desktop_mode_init">桌面模式</string>
-    <string name="about_title">关于</string>
+    <string name="home_title">关于</string>
     <string name="about_description">关于 MiceWine</string>
     <string name="general_settings">常规设置</string>
     <string name="exit">退出</string>
@@ -51,7 +51,7 @@
     <string name="env_edit_action">编辑环境变量。</string>
     <string name="env_key">名称</string>
     <string name="env_value">值</string>
-    <string name="home_title">主页</string>
+    <string name="shortcuts_title">主页</string>
     <string name="box64_settings_title">Box64 设置</string>
     <string name="box64_settings_description">配置 Box64</string>
